Alpha-30

+++ common/src/main/java/banduty/stoneycore/util/definitionsloader/AccessoriesDefinitionData.java
@@ -6,13 +6,13 @@ import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.EquipmentSlot;
 
 public record AccessoriesDefinitionData(double armor, double toughness, String armorSlot, double hungerDrainMultiplier,
-                                        double deflectChance, double weight, ResourceLocation visoredHelmet) {
+                                        float deflectChance, double weight, ResourceLocation visoredHelmet) {
     public static final Codec<AccessoriesDefinitionData> CODEC = RecordCodecBuilder.create(instance -> instance.group(
             Codec.DOUBLE.optionalFieldOf("armor", 0.0).forGetter(AccessoriesDefinitionData::armor),
             Codec.DOUBLE.optionalFieldOf("toughness", 0.0).forGetter(AccessoriesDefinitionData::toughness),
             Codec.STRING.optionalFieldOf("armorSlot", "").forGetter(AccessoriesDefinitionData::armorSlot),
             Codec.DOUBLE.optionalFieldOf("hungerDrainMultiplier", 0.0).forGetter(AccessoriesDefinitionData::hungerDrainMultiplier),
-            Codec.DOUBLE.optionalFieldOf("deflectChance", 0.0).forGetter(AccessoriesDefinitionData::deflectChance),
+            Codec.FLOAT.optionalFieldOf("deflectChance", 0.0f).forGetter(AccessoriesDefinitionData::deflectChance),
             Codec.DOUBLE.optionalFieldOf("weight", 0.0).forGetter(AccessoriesDefinitionData::weight),
             ResourceLocation.CODEC.optionalFieldOf("visoredHelmet", new ResourceLocation("", "")).forGetter(AccessoriesDefinitionData::visoredHelmet)
     ).apply(instance, AccessoriesDefinitionData::new));
